third_party/blink/web_tests/accessibility/idref-newlines.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div class="container">
  <input
    id="chk1"
    type="checkbox"
    aria-labelledby="label-1
                     label-2">
  <span id="label-1">nacho</span>
  <span id="label-2">cheese</span>
</div>

<script>
test(function(t) {
  var axCheckbox = accessibilityController.accessibleElementById("chk1");
  assert_equals(axCheckbox.name, "nacho cheese");
});
</script>
